ScaledIcon: render disabled state, and cache rendered images
It turns out the method that renders buttons disabled,
LookAndFeel.getDisabledIcon, requires an ImageIcon (despite taking
Icon as an argument). In other cases it returns null.

Therefore this class (an Icon that merely delegates to ImageIcon)
must take over the responsibility of rendering the disabled state itself,
but only if rendering onto a button.

Caching is added for both enabled & disabled states of *ImageIcon*s.
Rendering is already quite computationally intensive, and I was
already wanting to add caching - even more so now we might be doing
both high-quality rescaling and colour transforms on every paint.

We take care not to cache non-static icons (those with rollover or
pressed effects, for example) which is why caching is currently
limited to ImageIcons.