Make JPanel keep its original font
There may be specific places in the code in which we want to use the
original, smaller, font. For example when rendering text rotated 90-deg
sub-pixel anti-aliasing does not apply in which case Tahoma still
renders much better. This gives us a place to fetch that from which will
still continue to work with or without this library.