Make table grid lines much fainter
OK this tweak is not strictly related to sizing and scaling - but it applies
nicely to multiple L&Fs to make them a bit more 'modern' looking.

Many modern UIs drop grid lines completely, but that might not always be
practical to apply globally to tables not designed that way.

Note that this will work on L&Fs that define 'Table.gridColor' and have no
effect on others that don't (like Nimbus).